+++ pythainlp/tools/__init__.py
@@ -6,6 +6,7 @@ __all__: list[str] = [
     "get_full_data_path",
     "get_pythainlp_data_path",
     "get_pythainlp_path",
+    "is_unsafe_pickle_allowed",
     "is_offline_mode",
     "is_read_only_mode",
     "safe_print",
@@ -18,6 +19,7 @@ from pythainlp.tools.path import (
     get_full_data_path,
     get_pythainlp_data_path,
     get_pythainlp_path,
+    is_unsafe_pickle_allowed,
     is_offline_mode,
     is_read_only_mode,
 )